requirements.txt
numpy>=1.18.5
opencv-python >= 4.1
pillow
pytesseract
matplotlib>=3.2.2
#torch>=1.7.0
#torchvision>=0.8.1
torch=1.10.1
torchvision=0.11.2
pandas
requests
tqdm>=4.41.0
PyYAML>=5.3.1
scipy>=1.4.1
unidecode
dramatiq[redis, watch]
psutil